+++ src/click/_winconsole.py
@@ -6,6 +6,8 @@
 # compared to the original patches as we do not need to patch
 # the entire interpreter but just work in our little world of
 # echo and prompt.
+from __future__ import annotations
+
 import io
 import sys
 import time