+++ src/click/formatting.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import typing as t
 from contextlib import contextmanager
 from gettext import gettext as _