+++ tests/test_imports.py
@@ -27,6 +27,7 @@ click.echo(json.dumps(rv))
 """
 
 ALLOWED_IMPORTS = {
+    "__future__",
     "weakref",
     "os",
     "struct",